+++ src/lib.rs
@@ -201,9 +201,10 @@ pub async fn development_transport(
     keypair: identity::Keypair,
 ) -> std::io::Result<core::transport::Boxed<(PeerId, core::muxing::StreamMuxerBox)>> {
     let transport = {
-        let tcp = tcp::TcpConfig::new().nodelay(true);
-        let dns_tcp = dns::DnsConfig::system(tcp).await?;
-        let ws_dns_tcp = websocket::WsConfig::new(dns_tcp.clone());
+        let dns_tcp = dns::DnsConfig::system(tcp::TcpConfig::new().nodelay(true)).await?;
+        let ws_dns_tcp = websocket::WsConfig::new(
+            dns::DnsConfig::system(tcp::TcpConfig::new().nodelay(true)).await?,
+        );
         dns_tcp.or_transport(ws_dns_tcp)
     };
 
@@ -258,9 +259,10 @@ pub fn tokio_development_transport(
     keypair: identity::Keypair,
 ) -> std::io::Result<core::transport::Boxed<(PeerId, core::muxing::StreamMuxerBox)>> {
     let transport = {
-        let tcp = tcp::TokioTcpConfig::new().nodelay(true);
-        let dns_tcp = dns::TokioDnsConfig::system(tcp)?;
-        let ws_dns_tcp = websocket::WsConfig::new(dns_tcp.clone());
+        let dns_tcp = dns::TokioDnsConfig::system(tcp::TokioTcpConfig::new().nodelay(true))?;
+        let ws_dns_tcp = websocket::WsConfig::new(dns::TokioDnsConfig::system(
+            tcp::TokioTcpConfig::new().nodelay(true),
+        )?);
         dns_tcp.or_transport(ws_dns_tcp)
     };