+++ transports/dns/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Update to `libp2p-core` `v0.33.0`.
 
+- Remove implementation of `Clone` on `GenDnsConfig`. See [PR 2682].
+
+[PR 2682]: https://github.com/libp2p/rust-libp2p/pull/2682
+
 # 0.32.1
 
 - Update to `trust-dns` `v0.21`. See [PR 2543].